Fixes for different cameras usage
:Release Notes:
fixes to use cambufferlib for different cameras

:Detailed Notes:
usage of int* instead of size_t* in ReadData arg
linking librt.so to avoid linking errors
oss-pkg-info.yaml license file inclusion

:Testing Performed:
[QWTC-8085] [WebRTC] Verify WebEx call join using
            meeting URL
[QWTC-7943] [WebRTC] Verify MS Teams call join using
            meeting URL
build locally and tested in RPI4

:QA Notes:
NA

:Issues Addressed:
[WRO-12772] Stabilization of camshm new shared lib

Reviewed-on: http://gpro.lge.com/c/webosose/cambufferlib/+/336845
Tested-by: <moorthy.bs@lge.com>

+++ include/private/ISharedMemory.h
@@ -34,7 +34,7 @@ public:
     virtual bool Open(key_t shmemKey) = 0;
     virtual bool Create(key_t* shmemKey, const int unitSize, const int units) = 0;
     virtual bool Close() = 0;
-    virtual bool ReadData(uint8_t** buffer, size_t* len) = 0;
+    virtual bool ReadData(uint8_t** buffer, int* len) = 0;
     virtual bool WriteData(uint8_t* buffer, const size_t len) = 0;
 
 protected:
